feat(decision-perspective): unify gate execution logic under evaluatePerspectiveGate
